feat: rely on local semantic release for action
Remove the settings to install versions of semantic release, plugins,
and extensions. Instead, rely on any repo using this action on having
an installation of semantic release already in their repo, along with
configuration.

BREAKING CHANGE: Removed the following inputs to the action:
`semantic_version`, `branches`, `branch`, `extra_plugins`, `extends`,
and `working_directory`. The configuration should all live in the repo
in a `.releaserc` file.